fix bug in package deploy script, update deploy-base target to only nuke authorized_keys in prod, update motd

+++ Makefile
@@ -93,8 +93,8 @@ deploy-base: # setup newly-deployed target EC2 system to be ready for producing 
 	ssh $(EC2_USER)@$(EC2_HOST) "sudo mv -f /tmp/motd /etc/motd"
 	# run base build
 	ssh $(EC2_USER)@$(EC2_HOST) "bash --login -c 'sudo /home/ec2-user/bin/evon-deploy -b'"
-	# nuke the ssh pub key
-	ssh $(EC2_USER)@$(EC2_HOST) "bash --login -c 'sudo rm -f /home/ec2-user/.ssh/authorized_keys /root/.ssh/authorized_keys'"
+	# nuke the ssh pub key if prod, ready for AMI creation (to pass AWS MP security scan)
+	[ "$(ENV)" == "prod" ] && ssh $(EC2_USER)@$(EC2_HOST) "bash --login -c 'sudo rm -f /home/ec2-user/.ssh/authorized_keys /root/.ssh/authorized_keys'" || :
 	echo Done.
 
 deploy-test: # DEV ONLY - convenience target to make package, publish and run installer on remote host
